requirements.txt

amqp==2.4.1
astroid==2.1.0
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
bleach==3.1.0
blessings==1.7
boto3==1.9.113
botocore==1.12.113
bpython==0.17.1
bs4==0.0.1
celery==4.2.1
certifi==2018.10.15
cffi==1.12.0
chardet==3.0.4
colorama==0.4.1
curtsies==0.3.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-autoslug==1.9.3
django-celery-results==1.0.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.4
django-haystack==2.8.1
django-heroku==0.3.1
django-mathfilters==0.4.0
django-mysql==2.4.1
django-pdb==0.6.2
django-shortcuts==1.6
django-silk==3.0.1
django-storages==1.7.1
django-tables2==2.0.3
docopt==0.6.2
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
furl==2.0.0
future==0.17.1
geordi==0.3
gevent==1.4.0
gprof2dot==2016.10.13
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipdb==0.12
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
kombu==4.3.0
lazy-object-proxy==1.3.1
lxml==4.3.2
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
metakernel==0.20.14
mistune==0.8.4
mock==2.0.0
mysql-connector-python==8.0.15
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.6
num2words==0.5.9
numpy==1.16.2
octave-kernel==0.28.4
openpyxl==2.6.1
orderedmultidict==1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.7
ptyprocess==0.6.0
pycodestyle==2.4.0
pycparser==2.19
Pygments==2.3.0
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
PyMySQL==0.9.3
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.7.5
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
pytz==2018.7
pywinpty==0.5.5
pyzmq==18.0.1
requests==2.20.1
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.11.0
soupsieve==1.8
South==1.0.2
spotipy==2.4.4
SQLAlchemy==1.2.17
sqlparse==0.2.4
tabulate==0.8.3
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
tqdm==4.30.0
traitlets==4.3.2
typing==3.6.6
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
wrapt==1.10.11
XlsxWriter==1.1.5
